Fix to post to archive queue whenever Frame object is updated
Prior to this, the archive queue post would only happen on frame creation. This becomes an issue if a thumbnail makes it to the archive before the FITS file - we cannot post to queue consumers until after the FITS image is ingested, so make sure we do that in the serializer rather than the view.